Add typescript compiler and setup watch commands
