--- src/main.js
@@ -1,14 +1,0 @@
-class TypesetHyphen {
-    line = "hello world";
-  
-    test = function() {
-      console.log('test2');
-    }
-  }
-  
-class Typeset extends TypesetHyphen {
-other = [1, 2, 3];
-test2 = function () {
-    console.log('test');
-};
-}
