+++ src/ts/tsb.hyphen.ts
@@ -1,0 +1,10 @@
+class TypesetHyphen {
+    line : string = "hello world";
+
+    test = function() {
+        console.log('test2');
+    };
+    hello = function() {
+        return 10;
+    } 
+}
