+++ src/ts/tsb.ts
@@ -1,0 +1,7 @@
+class Typeset extends TypesetHyphen {
+    other = [1, 2, 3];
+    test2 = function () {
+        console.log('test');
+      };
+  }
+
